Increased Puzzle Image Size
To support zooming, the puzzle image is rendered at a higher quality than is typically needed. I noticed that game images were still looking quite blurry when zoomed in, so this change increases the size of the image to be 2048px on square (up from 1024). If this causes performance issues on some devices, I may need to consider making this a setting.